Super R-Type and Gradius III: Enable I-RAM write before using stack operations
At power and reset, I-RAM must be write protected. Looks like this is not 100% strictly emulated.

Both games attempts writing setting up the data bank (PLB), which relies on the stack, which relies on the I-RAM, without enabling I-RAM write. Because of that, some cartridges (yet unknown why some only is affected) won't write to I-RAM and when using PLB, the previous value is read (which depending on the chip, either it will return a random value from I-RAM or the previous stored value), setting the wrong data bank and likely crashing the game on a second moment.